Use size_t as variable type for storage of sizeof(staticbuf) (cherry picked from commit 03223ed5eb357c8d9c35c210a742d5cbd2644af4)
Conflicts:
	src/common/image.c
	src/common/image.h

+++ src/libs/recentcollect.c
@@ -99,7 +99,7 @@ void connect_key_accels(dt_lib_module_t *self)
 
 
 static void
-pretty_print(char *buf, char *out, int outsize)
+pretty_print(char *buf, char *out, size_t outsize)
 {
   memset(out, 0, outsize);
